feat: add LanguageSwitch component and integrate it into Header and LoginPage
