+++ frontend/src/components/layout/Header.tsx
@@ -1,8 +1,8 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import ThemeSwitch from '@/components/ui/ThemeSwitch';
+import LanguageSwitch from '@/components/ui/LanguageSwitch';
 import GitHubIcon from '@/components/icons/GitHubIcon';
-import LanguageIcon from '@/components/icons/LanguageIcon';
 import SponsorDialog from '@/components/ui/SponsorDialog';
 import WeChatDialog from '@/components/ui/WeChatDialog';
 
@@ -11,40 +11,9 @@ interface HeaderProps {
 }
 
 const Header: React.FC<HeaderProps> = ({ onToggleSidebar }) => {
-  const { t, i18n } = useTranslation();
+  const { t } = useTranslation();
   const [sponsorDialogOpen, setSponsorDialogOpen] = useState(false);
   const [wechatDialogOpen, setWechatDialogOpen] = useState(false);
-  const [languageDropdownOpen, setLanguageDropdownOpen] = useState(false);
-  const [currentLanguage, setCurrentLanguage] = useState(i18n.language);
-
-  // Update current language when it changes
-  useEffect(() => {
-    setCurrentLanguage(i18n.language);
-  }, [i18n.language]);
-
-  // Close dropdown when clicking outside
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      const target = event.target as HTMLElement;
-      if (!target.closest('.language-dropdown')) {
-        setLanguageDropdownOpen(false);
-      }
-    };
-
-    if (languageDropdownOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [languageDropdownOpen]);
-
-  const handleLanguageChange = (lang: string) => {
-    localStorage.setItem('i18nextLng', lang);
-    setLanguageDropdownOpen(false);
-    window.location.reload();
-  };
 
   return (
     <header className="bg-white dark:bg-gray-800 shadow-sm z-10">
@@ -109,41 +78,7 @@ const Header: React.FC<HeaderProps> = ({ onToggleSidebar }) => {
             <SponsorIcon className="h-5 w-5" />
           </button> */}
           <ThemeSwitch />
-          {/* Language Switcher */}
-          <div className="relative language-dropdown">
-            <button
-              onClick={() => setLanguageDropdownOpen(!languageDropdownOpen)}
-              className="p-2 rounded-md text-gray-500 dark:text-gray-400 hover:text-gray-900 dark:hover:text-gray-200 hover:bg-gray-100 dark:hover:bg-gray-700 focus:outline-none"
-              aria-label="Language Switcher"
-            >
-              <LanguageIcon className="h-5 w-5" />
-            </button>
-
-            {languageDropdownOpen && (
-              <div className="absolute right-0 mt-2 w-24 bg-white dark:bg-gray-800 rounded-md shadow-lg border border-gray-200 dark:border-gray-700 z-50">
-                <div className="py-1">
-                  <button
-                    onClick={() => handleLanguageChange('en')}
-                    className={`flex items-center w-full px-4 py-2 text-sm hover:bg-gray-100 dark:hover:bg-gray-700 ${currentLanguage.startsWith('en')
-                      ? 'text-blue-600 dark:text-blue-400 bg-blue-50 dark:bg-blue-900/20'
-                      : 'text-gray-700 dark:text-gray-300'
-                      }`}
-                  >
-                    English
-                  </button>
-                  <button
-                    onClick={() => handleLanguageChange('zh')}
-                    className={`flex items-center w-full px-4 py-2 text-sm hover:bg-gray-100 dark:hover:bg-gray-700 ${currentLanguage.startsWith('zh')
-                      ? 'text-blue-600 dark:text-blue-400 bg-blue-50 dark:bg-blue-900/20'
-                      : 'text-gray-700 dark:text-gray-300'
-                      }`}
-                  >
-                    中文
-                  </button>
-                </div>
-              </div>
-            )}
-          </div>
+          <LanguageSwitch />
         </div>
       </div>
       <SponsorDialog open={sponsorDialogOpen} onOpenChange={setSponsorDialogOpen} />
